command line arguments for client and server implemented, Folders restructured
